% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumnTypicalDay.R
\name{addTypicalDayId}
\alias{addTypicalDayId}
\title{Add a flow-based typical day id column}
\usage{
addTypicalDayId(data, fb_opts = antaresRead::simOptions())
}
\arguments{
\item{data}{\code{antaresdata} output data load by \link{readAntares}}

\item{fb_opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath} or fb model 
localisation obtain with \link{setFlowbasedPath}. 
The default value is indicated by \code{antaresRead::simOptions()}.}
}
\description{
Add to an Antares output table a column indicating the id of the used flow-based typical day for each timestep.
This information is read in the "user" directory file of the Antares study: it must be made sure that the data of the study is still
consistent with the output.
}
\examples{

\dontrun{
study <- "../../Pour Julien/blop/MT_base_nucM2_2023"
antaresRead::setSimulationPath(study, 17)
data <- readAntares(mcYears = 1:2)
data <- addTypicalDayId(data)


data <- readAntares(areas = "all", links = "all", clusters = "all" ,mcYears = 1:2)
data <- addTypicalDayId(data)

}

}
