% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFBTS.R
\name{createFBTS}
\alias{createFBTS}
\title{Creation of time series of typical days for Antares studies}
\usage{
createFBTS(opts, probabilityMatrix, multiplier, firstDay, seed = 123456,
  silent = FALSE, outputPath = getwd(),
  calendar = system.file("calendar/calendar.txt", package = "fbAntares"))
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function 
\link{setSimulationPath}. Link to the Antares study.
 By default, the value is \code{antaresRead::simOptions()}}

\item{probabilityMatrix}{\code{list}, correlation between climatic factors and 
flow-based typical days, such as returned by \code{fbClust::getProbability}. 
The columns names must be renamed to match Antares' inputs, use the function
 \link{setNamesProbabilityMatrix}. Initial format is : 
 area_variable (Ex: fr_load, de_solar ...)}

\item{multiplier}{\code{data.frame} enabling to convert load factors or normalised 
values into production/consumption in MW. 
Two columns:
\itemize{
\item variable : Name of variable (ex: \code{"fr@wind"})
\item coef : mutiplier coefficient, for example the installed capacity.
}}

\item{firstDay}{\code{numeric} Type of the first day of the study (between 1 and 7). 
For example, if the first day is a
Wednesday, you must specify firstDay = 3. The first day can be directly calculated
by the function \link{identifyFirstDay}.}

\item{seed}{\code{numeric} fixed random seed, used for the weighted draw of the 
typical days. By default, the value is 123456}

\item{silent}{\code{boolean}, non display of a progress bar. By default, the value is FALSE.}

\item{outputPath}{\code{character}, path of the folder where the time series of 
typical flow-based output file (ts.txt) will be written. The current directory 
is chosen by default.}

\item{calendar}{\code{character}, path to a txt files with two columns,
a column Date in the format month-day and a column class with the classes
associated to the dates (summer, winter and interSeason).}
}
\description{
This function defines for each studied day in an Antares study the most 
representative typical flow-based day, and finally creates daily time series.
To establish this correlation, the inputs of the function must include 
a probability matrix (calculated for each set of typical days with the package 
\code{fbClust}) and the path to the Antares study to provide flow-based 
domains with. The probability matrix
will be used to compute a weighted draw among the possible typical days.
}
\examples{

\dontrun{
library(antaresRead)
library(data.table)

matProb <- readRDS(system.file("testdata/MatProb.rds", package = "fbAntares"))

# Set the probabilityMatrix names and coefficients
matProb <- setNamesProbabilityMatrix(matProb, c("FR_load", "DE_wind", "DE_solar"),
                                   c("fr@load", "de@wind", "de@solar"))

multiplier <- data.frame(variable = c("fr@load", "de@wind", "de@solar"),
                         coef = c(1, 71900, 61900))

# Set the path to Antares study inputs 
# Change it for the path of a study you have on your computer
opts <- antaresRead::setSimulationPath("../../Pour Julien/blop/MT_base_nucM2_2023/", 
simulation = 17)

# calendar
# first day identified based on the input data of the 
# Antares study designated by opts
firstDay <- identifyFirstDay(opts) 
# calendar in a csv file with all dates of one year and their classes
calendar <- system.file("calendar/calendar.txt", package = "fbAntares")

# Generate flow-based time series
ts <- createFBTS(opts = opts, probabilityMatrix = matProb, multiplier = multiplier,
                 calendar = calendar, firstDay = firstDay, outputPath = getwd())
}
                 
}
