% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{drawVolumeHeatmapThreeDimensions}
\alias{drawVolumeHeatmapThreeDimensions}
\title{Draw a volume heatmap comparing the volumes of two polyhedrons making combinations
over the range of three possible net positions}
\usage{
drawVolumeHeatmapThreeDimensions(
  A,
  B = NULL,
  nbPoints = 50000,
  seed = 123456,
  assessment_range = c(-15000, 15000),
  ptdf_list
)
}
\arguments{
\item{A}{\code{data.table}, polyhedron, data.table containing at least 
two ptdf columns :
\itemize{
 \item ptdfAT : autrichian vertices
 \item ptdfBE : belgium vertices
 \item ptdfDE : german vertices
 \item ptdfFR : french vertices
}}

\item{B}{\code{data.table}, polyhedron, data.table containing at least 
two ptdf columns :
\itemize{
 \item ptdfAT : autrichian vertices
 \item ptdfBE : belgium vertices
 \item ptdfDE : german vertices
 \item ptdfFR : french vertices
}
If B is NULL, the heatmap returns the absolute volume of A in the different directions.}

\item{nbPoints}{\code{numeric}, number of points generated for volume comparison}

\item{seed}{\code{numeric} fixed random seed, used for the weighted draw of the
points for volume assessment. By default, the value is 123456}

\item{assessment_range}{The range in which the points for the volume assessment should be drawn}

\item{ptdf_list}{The list of ptdfs over which the heatmap must be drawn. They are names of columns
of A and B.}
}
\description{
This function returns a 3D heatmap comparing the volumes of two polyhedrons
in different directions corresponding to all the combinations of the positive/negative
net positions of each of the three countries provided in the \code{ptdf_list} argument.
}
\examples{
\dontrun{
library(data.table)
polyhedra <- readRDS(system.file("testdata/polyhedra.rds", package = "fbAntares"))
A <- polyhedra[Date == "2019-02-14"]
B <- polyhedra[Date == "2019-02-15"]
nbPoints <- 50000

 drawVolumeHeatmapThreeDimensions(A = A, B = B, nbPoints = nbPoints, ptdf_list = c("ptdfFR", "ptdfBE", "ptdfDE"))
}

}
