% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{drawVolumeHeatmapTwoDimensions}
\alias{drawVolumeHeatmapTwoDimensions}
\title{Draw a volume heatmap comparing the volumes of two polyhedrons making combinations
of possible net positions for the studied countries}
\usage{
drawVolumeHeatmapTwoDimensions(
  A,
  B = NULL,
  nbPoints = 50000,
  seed = 123456,
  assessment_range = c(-15000, 15000)
)
}
\arguments{
\item{A}{\code{data.table}, polyhedron, data.table containing at least 
two ptdf columns :
\itemize{
 \item ptdfAT : autrichian vertices
 \item ptdfBE : belgium vertices
 \item ptdfDE : german vertices
 \item ptdfFR : french vertices
}}

\item{B}{\code{data.table}, polyhedron, data.table containing at least 
two ptdf columns :
\itemize{
 \item ptdfAT : autrichian vertices
 \item ptdfBE : belgium vertices
 \item ptdfDE : german vertices
 \item ptdfFR : french vertices
}
If B is NULL, the heatmap returns the absolute volume of A in the different directions.}

\item{nbPoints}{\code{numeric}, number of points generated for volume comparison}

\item{seed}{\code{numeric} fixed random seed, used for the weighted draw of the
points for volume assessment. By default, the value is 123456}

\item{assessment_range}{The range in which the points for the volume assessment should be drawn}
}
\description{
This function returns a heatmap comparing the volumes of two polyhedrons
in different directions corresponding to all the combinations of the positive/negative
net positions of each country
}
\examples{
\dontrun{
library(data.table)
polyhedra <- readRDS(system.file("testdata/polyhedra.rds", package = "fbAntares"))
A <- polyhedra[Date == "2019-02-14"]
B <- polyhedra[Date == "2019-02-15"]
nbPoints <- 50000

 drawVolumeHeatmapTwoDimensions(A = A, B = B, nbPoints = nbPoints)
}

}
