% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntersectionLinesHyperplanes.R
\name{evalDomainVolume}
\alias{evalDomainVolume}
\title{Evaluate the volume of a polyhedron}
\usage{
evalDomainVolume(
  A,
  nbPoints = 50000,
  seed = 123456,
  assessment_range = c(-15000, 15000),
  direction = NULL
)
}
\arguments{
\item{nbPoints}{\code{numeric}, number of points generated}

\item{seed}{\code{numeric} fixed random seed, used for the weighted draw of the 
points. By default, the value is 123456}

\item{assessment_range}{The range in which the points for the volume assessment should be drawn.
This range should cover the entire domain described by A}

\item{direction}{\code{data.table} used to provide specific directions in which
the volume calculation should be performed. The data table must contain two columns:
\itemize{
 \item a column "country" containing the names of the countries on which a direction is specified
 \item a column "direction" containing a string which can be either "positive" to specify that only
 positive net positions must be considered for this country, or "negative" if only negative net
 positions must be considered
}}
}
\value{
The volume of polyhedron A
}
\description{
This function returns the volume of a polyhedron by calling the function
\code{evalInter} in order to find the shared volume between this polyhedron and a standard
polyhedron which dimension is known.
}
\examples{
\dontrun{
library(data.table)
polyhedra <- readRDS(system.file("testdata/polyhedra.rds", package = "fbAntares"))
A <- polyhedra[Date == "2019-02-14"]
nbPoints <- 50000

 evalDomainVolume(A = A, nbPoints = nbPoints)
}

}
