% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntersectionLinesHyperplanes.R
\name{evalInter}
\alias{evalInter}
\title{Evaluate the shared volume of two polyhedra A and B}
\usage{
evalInter(A, B, nbPoints = 50000)
}
\arguments{
\item{A}{\code{data.table}, fix polyhedron, data.table containing at least 
two ptdf columns :
\itemize{
 \item ptdfAT : autrichian vertices
 \item ptdfBE : belgium vertices
 \item ptdfDE : german vertices
 \item ptdfFR : french vertices
}}

\item{nbPoints}{\code{numeric}, number of points generated}

\item{A}{\code{data.table}, moving polyhedron, data.table containing at least 
two ptdf columns :
\itemize{
 \item ptdfAT : autrichian vertices
 \item ptdfBE : belgium vertices
 \item ptdfDE : german vertices
 \item ptdfFR : french vertices
}}
}
\description{
This function returns and indicator between 0 and 1 where 0 means
the intersection between A and B is empty and 1 means A == B.
Since there is no fast metric to compute the volume of a polyhedron in high 
dimension, this indicator is computed by generating n points in a n-dimensonal
space doing the ratio of the points in the two polyhedra and the points in
at least one of the polyhedra.
}
\examples{
\dontrun{
library(data.table)
polyhedra <- readRDS("~/RTE/fbAntares/inst/testdata/polyhedra.rds")
A <- polyhedra[Date == "2019-02-14"]
B <- polyhedra[Date == "2019-02-15"]
nbPoints <- 50000

 evalInter(A = A, B = B, nbPoints = nbPoints)
}

}
