% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFlowbasedPath.R
\name{flowbased-path}
\alias{flowbased-path}
\alias{setFlowbasedPath}
\alias{fbOptions}
\alias{getAvailableModel}
\title{Set the flow-based input folder}
\usage{
setFlowbasedPath(path, model)

fbOptions()

getAvailableModel()
}
\arguments{
\item{path}{(optional)
If the argument "model" is missing. Path to the flow-based input directory, 
it can be created by the user with the functions \link{computeFB} and 
\link{createFBTS}. The directory must be composed of:
\itemize{
 \item{weight.txt}{ the PTDF file of the Antares models}
 \item{second_member.txt}{ the RAM of the Antares model}
 \item{ts.txt}{ the flow-based time series}
 \item{domainesFB.RDS}{ information about the domains}
}}

\item{model}{(optional) 
If "path" is missing. Name of the chosen existing model 
(Use \code{\link{getAvailableModel}} to get the complete list).}
}
\value{
\itemize{ 
\item {}{A vector of available models for \code{getAvailableModel}.}
\item {}{path of the current input repository, for \code{setFlowbasedPath}
 and \code{fbOptions}}
}
}
\description{
A flow-based "model" consists of the description of the typical
 domains (the weights, the second members and an RDS object containing 
 additional information such as the vertices coordinates and the errors) and 
 the time series file establishing the chronology of typical days used in an Antares study. 
 These functions enable to choose the flow-based input folder: one can either use 
 existing models stored in the package or use his own. 
  \itemize{
   \item{The function \code{getAvailableModel} provides a list of the existing models.}
   \item{The function \code{setFlowbasedPath} sets the flow-based input folder. 
   It can either be one of the models listed by 
   \code{getAvailableModel} or a path to a personal folder. This folder will then 
   be used as default parameter for all
    functions of the package}
   \item{The function \code{fbOptions} indicates the current default path.}
  }
}
\examples{

\dontrun{
# Get the default path used when loading the package
fbOptions()

# Specify an available model
getAvailableModel()
setFlowbasedPath(model = "model2017")

# Specify a path
setFlowbasedPath(path = system.file("testdata/antaresInput", package = "fbAntares"))

# Specify a personnal model
setFlowbasedPath(model = "C:/PATH/TO/INPUT")
}

}
