% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxImportExport.R
\name{getMaxImportExport}
\alias{getMaxImportExport}
\title{Return a data.table with the maximum values of import and export}
\usage{
getMaxImportExport(domainesFB)
}
\arguments{
\item{domainesFB}{\code{data.table} The output domainesFB obtained with \link{computeFB}}
}
\description{
This function takes the object domainesFB, generated with \link{computeFB}
and return a data.table with the maximum values of import and export by
period and typical day.
}
\examples{
\dontrun{
domainesFB <- readRDS(system.file("testdata/antaresInput/domainesFB.rds", package = "fbAntares"))
dtImportExport <- getMaxImportExport(domainesFB)
}
}
