% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatsFBts.R
\name{getStatsFBts}
\alias{getStatsFBts}
\title{Return a list with statistics summary of the file ts.txt}
\usage{
getStatsFBts(ts, calendar = system.file("calendar/calendar.txt", package
  = "fbAntares"), output = "summary")
}
\arguments{
\item{ts}{\code{data.table} The output ts obtained with \link{createFBTS}}

\item{calendar}{\code{character}, path to a txt files with two columns,
a column Date in the format month-day and a column class with the classes
associated to the dates (summer, winter and interSeason).}

\item{output}{\code{character}, the output of you want, either summary, 
yearbyyear or all, default is summary, if you want the details check the example.}
}
\description{
This function takes the object ts, generated with \link{createFBTS}
and return a list with statistics summary of this object
}
\examples{
\dontrun{
ts <- fread(system.file("testdata/antaresInput/ts.txt", package = "fbAntares"),
header = T)
calendar <- system.file("calendar/calendar.txt", package = "fbAntares")
statsFBts <- getStatsFBts(ts, calendar, output = "summary")
statsFBts <- getStatsFBts(ts, calendar, output = "yearbyyear")
statsFBts <- getStatsFBts(ts, calendar, output = "all")
}
}
