% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEntsoeXML.R
\name{readEntsoeXML}
\alias{readEntsoeXML}
\title{Download data from entsoe and convert them to data.table and csv.}
\usage{
readEntsoeXML(request = NULL, bz = NULL, outDomain = NULL,
  docType = NULL, docStatus = NULL, timeStart = NULL,
  timeEnd = NULL, processType = NULL, psrType = NULL, token = NULL,
  entsoeHttps = "https://transparency.entsoe.eu/api?",
  writeDataDirectory = NULL, fileName = NULL)
}
\arguments{
\item{request}{\code{character} If you are familiar with entsoe data, you can
put here your entire request instead of use the decomposition proposed with
the other parameters}

\item{bz}{\code{character} build zone, transparency.entsoe.eu API guide}

\item{outDomain}{\code{character} outDomain (if there is one), an area
linked to the build zone}

\item{docType}{\code{character} docType, A80 for generation unit, A77 for 
produciton unit, A69 for solar and wind forecast,}

\item{docStatus}{\code{character} docStatus "A05" : Active, "A09" : Cancel}

\item{timeStart}{\code{character} start date, format YYYYMMDDhhmm (ex : 201807200000)}

\item{timeEnd}{\code{character} end date, format YYYYMMDDhhmm (ex : 201907200000)
Maximum of documents load by request is 200 so if data are not correctly try 
an other timestep.}

\item{processType}{\code{character} type of processed data you want, ex : 
A16 = realised, A33 = year ahead}

\item{psrType}{\code{character} type of energy you want, ex : B16 = Solar,
B19 = Wind Onshore}

\item{token}{\code{character}, ENTSOE token}

\item{entsoeHttps}{\code{character}, ENTSOE adress, defalut https://transparency.entsoe.eu/api?}

\item{writeDataDirectory}{\code{character} Write the data in a csv or rds file in
the wanted directory}

\item{fileName}{\code{character} Name of the file you want to write 
(either csv or rds). If NULL, the file is writtend in your working directory}
}
\description{
This function allows you to download data from entsoe and convert 
it in data.frame and/or a csv file. If you need more help to kwon how to do 
the requests, don't hesitate to use the help of the entsoe website :
https://transparency.entsoe.eu/content/static_content/Static%20content/web%20api/Guide.html#_available_parameters
Some of the following arguments can be optional according to the data you 
want to load.

This function works for all wind and solar data but can meet some issues 
with other data
}
\examples{
\dontrun{


 #PROD : "A77" Gene : "A80"
 token <- "Mytoken" 
 
 #BiddingZone_Domain = "10YFR-RTE------C"
 #BiddingZone_Domain = "10YBE----------2"
 #BiddingZone_Domain = "10YNL----------L"
 #BiddingZone_Domain = "10Y1001A1001A63L"

 #BiddingZone_Domain = "10YDE-VE-------2"    #GRT Allemand 50Hertz
 #BiddingZone_Domain = "10YDE-RWENET---I"    #GRT Allemand Amprion
 #BiddingZone_Domain = "10YDE-EON------1"    #GRT Allemand TennetDE
 #BiddingZone_Domain = "10YDE-ENBW-----N"    #GRT Allemand TransnetBW
 #BiddingZone_Domain = "10YAT-APG------L"    #Autriche#' 

 #NL Wind + Solar
 readEntsoeXML(token = token, bz = "10YNL----------L", docType = "A69",
start = "201807200000", end = "201907230000", processType = "A01")

 #BE Wind + Solar
 readEntsoeXML(token = token, bz = "10YBE----------2", docType = "A69",
start = "201807200000", end = "201907230000", processType = "A01")


 #FR Wind + Solar
 readEntsoeXML(token = token, bz = "10YFR-RTE------C", docType = "A69",
start = "201807200000", end = "201907230000", processType = "A01")
 
 
 readEntsoeXML(token = token, bz = "10YFR-RTE------C", docType = "A80",
   start = "2015-01-01", end = "2018-12-31", fileToMerge = "FRindispo.csv")
 readEntsoeXML(token = token, "10YFR-RTE------C", docType = "A77",
   start = "2015-01-01", end = "2018-12-31", fileToMerge = "FRindispo.csv")

 #DE wind solar some examples
 
 dtDESolarWind1 <- readEntsoeXML(bz = "10YDE-VE-------2", token = token,
docType = "A69", start = "201807200000", end = "201907230000",
processType = "A01")

 dtDESolarWind2 <- readEntsoeXML(bz = "10YDE-RWENET---I", token = token,
docType = "A69", start = "201807200000", end = "201907230000",
processType = "A01",  writeDataDirectory = getwd())

 dtDESolarWind3 <- readEntsoeXML(bz = "10YDE-EON------1", token = token,
docType = "A69", start = "201807200000", end = "201907230000",
processType = "A01")

 dtDESolarWind4 <- readEntsoeXML(bz = "10YDE-ENBW-----N", token = token,
docType = "A69", start = "201807200000", end = "201907230000",
processType = "A01", writeDataDirectory = getwd(),
fileName = "solarwindDEpart4.rds")

 dtDEWind1 <- readEntsoeXML(bz = "10YDE-VE-------2", token = token,
docType = "A69", start = "201807200000", end = "201907230000",
processType = "A01", psrType = "B19",  writeDataDirectory = tempdir(),
fileName = "windDEpart1.csv") 
   
 dtDESolar2 <- readEntsoeXML(bz = "10YDE-RWENET---I", token = token,
docType = "A69", start = "201807200000", end = "201907230000",
processType = "A01", psrType = "B16",  writeDataDirectory = getwd())
   
   
}

}
