% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{runAppError}
\alias{runAppError}
\title{Run a shiny application to visualize the real and modelled flow-based domains}
\usage{
runAppError(fb_opts = fbAntares::fbOptions(), countries = list(c("BE",
  "FR"), c("BE", "NL"), c("DE", "FR"), c("DE", "AT")))
}
\arguments{
\item{fb_opts}{\code{list} of flow-based parameters returned by the function 
\link{setFlowbasedPath}: directory of the flow-based model to study.
By default, the value is indicated by \code{fbAntares::fbOptions()}}

\item{countries}{\code{list, character} a list of couples of countries to choose the axises for the projection
of the flowbased domains (ex : list(c("BE", "FR"), c("BE", "NL"))) or an array of countries
(ex : c("FR", "NL", "AT")) to project the domains on all the countries combination 
(here FR+NL, FR+AT and NL+AT)}
}
\description{
Run a shiny application displaying the results of the conversion of real domains into Antares models. It will display for
each typical day and each hour the volumetric errors of conversion (inf_error: forgotten points in the 
model, sup_error: modelled points missing in the real domain) and dynamic plots of the real and modelled domains. Html reports
for each day can also be exported.
}
\examples{

\dontrun{
 fb_opts = fbAntares::fbOptions()
 runAppError(fb_opts, 
 countries = list(c("BE", "FR"), c("BE", "NL"), c("DE", "FR"), c("DE", "AT")))
}
}
