% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFBTS.R
\name{setNamesProbabilityMatrix}
\alias{setNamesProbabilityMatrix}
\title{Rename the probabilityMatrix variables}
\usage{
setNamesProbabilityMatrix(data, oldName, newName)
}
\arguments{
\item{data}{\code{list}, probabilityMatrix whose columns to rename,
calculated with \code{flowBasedClustering::getProbability}.}

\item{oldName}{\code{character} vector of old variables names}

\item{newName}{\code{character} vector of new variables names}
}
\description{
This function changes the name of the variables in the probabilityMatrix, 
created by \code{flowBasedClustering::getProbability} and 
used by \link{createFBTS}. Aim: rename the variables into a format consistent 
with Antares inputs ("area@variable"). For example, go to \link{createFBTS}.
}
