% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFBTS.R
\name{createFBTS}
\alias{createFBTS}
\title{Creation of time series of typical days for Antares studies}
\usage{
createFBTS(opts, probabilityMatrix, multiplier, interSeasonBegin,
  interSeasonEnd, firstDay, seed = NULL, silent = FALSE,
  outputPath = getwd())
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function 
\link{setSimulationPath}. Link to the Antares study.
 By default, the value is \code{antaresRead::simOptions()}}

\item{probabilityMatrix}{\code{list}, correlation between climatic factors and 
flow-based typical days, such as returned by \code{flowBasedClustering::getProbability}. 
The columns names must be renamed to match Antares' inputs, use the function
 \link{setNamesProbabilityMatrix}. Initial format is : area_variable (Ex: fr_load, de_solar ...)}

\item{multiplier}{\code{data.frame} enabling to convert load factors or normalised 
values into production/consumption in MW. 
Two columns:
\itemize{
\item variable : Name of variable (ex: \code{"fr@wind"})
\item coef : mutiplier coefficient, for example the installed capacity.
}}

\item{interSeasonBegin}{\code{character or date}, date or vector of dates, YYYY-MM-DD, 
begin of interseason}

\item{interSeasonEnd}{\code{character or date}, date or vector of dates, YYYY-MM-DD, 
end of interseason}

\item{firstDay}{\code{numeric} Type of the first day of the study (between 1 and 7). 
For example, if the first day is a
Wednesday, you must specify firstDay = 3. The first day can be directly calculated
by the function \link{identifyFirstDay}.}

\item{seed}{\code{numeric} fixed random seed, used for the weighted draw of the 
typical days. By default, the value is NULL.}

\item{silent}{\code{boolean}, non display of a progress bar. By default, the value is FALSE.}

\item{outputPath}{\code{character}, path of the folder where the time series of 
typical flow-based output file (ts.txt) will be written. The current directory is chosen by default.}
}
\description{
This function defines for each studied day in an Antares study the most 
representative typical flow-based day, and finally creates daily time series.
To establish this correlation, the inputs of the function must include 
a probability matrix (calculated for each set of typical days with the package 
\code{flowBasedClustering}) and the path to the Antares study to provide flow-based 
domains with. The probability matrix
will be used to compute a weighted draw among the possible typical days.
}
\examples{

\dontrun{
library(antaresRead)
library(data.table)

matProb <- readRDS(system.file("testdata/MatProb.rds", package = "fbAntares"))

# Set the probabilityMatrix names and coefficients
matProb <- setNamesProbabilityMatrix(matProb, c("FR_load", "DE_wind", "DE_solar"),
                                   c("fr@load", "de@wind", "de@solar"))

multiplier <- data.frame(variable = c("fr@load", "de@wind", "de@solar"),
                         coef = c(1, 352250, 246403))

# Set the path to Antares study inputs 
opts <- antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/antaresStudy", 1)

# calendar
# first day identified based on the input data of the 
# Antares study designated by opts
firstDay <- identifyFirstDay(opts) 
interSeasonBegin <- as.Date(c("2017-09-03", "2018-02-02"))
interSeasonEnd <- as.Date(c("2020-10-04", "2018-05-02"))

# Generate flow-based time series
ts <- createFBTS(opts = opts, probabilityMatrix = matProb, multiplier = multiplier,
                 interSeasonBegin = interSeasonBegin,
                 interSeasonEnd = interSeasonEnd, firstDay = firstDay, outputPath = getwd())
}
                 
}
