% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDist.R
\name{getBestPolyhedron}
\alias{getBestPolyhedron}
\title{Adjust a polyhedron from a set of facets to approach a reference polyhedron}
\usage{
getBestPolyhedron(A, B, nbLines, maxiter, thresholdIndic, quad = F,
  verbose = 2, seed = 123456, fixFaces, VERTRawDetails)
}
\arguments{
\item{A}{\code{data.table}, fix polyhedron, data.table containing at least 
ram, Date, Period and two ptdf columns :
\itemize{
 \item ptdfAT : autrichian ptdf
 \item ptdfBE : belgium ptdf
 \item ptdfDE : german ptdf
 \item ptdfFR : french ptdf
 \item ram : line limits
 \item Date : date in format YYYY-MM-DD
 \item Period : hour in the day, between 1 and 24
}}

\item{B}{\code{data.table}, moving polyhedron, data.table containing at least 
ram, Date, Period and two ptdf columns :
\itemize{
 \item ptdfAT : autrichian ptdf
 \item ptdfBE : belgium ptdf
 \item ptdfDE : german ptdf
 \item ptdfFR : french ptdf
 \item ram : line limits
 \item Date : date in format YYYY-MM-DD
 \item Period : hour in the day, between 1 and 24
}}

\item{nbLines}{\code{numeric}, number of half-lines used to compute the distance
between the two polyhedra.}

\item{maxiter}{\code{numeric}, maximum number of iteration, it is one of the 
two stopping criteria of this function.}

\item{thresholdIndic}{\code{numeric}, maximum number of iteration, it is one of the 
two stopping criteria of this function.}

\item{quad}{\code{logical}, TRUE if you want to solve it with a quadratic
optimization problem, FALSE if you want to use a linear (default is linear, 
which is faster)}

\item{verbose}{\code{numeric}, shows the logs in console. By default, the value is 1.
\itemize{
 \item 0 : No log
 \item 1 : No log
 \item 2 : Medium log
 }}

\item{seed}{\code{numeric} fixed random seed, used for the weighted draw of the 
typical days. By default, the value is 123456}

\item{fixFaces}{\code{data.table} data.table if you want to use fix faces for the creation
of the flowbased models. If you want to do it, the data.table has the following form :
data.table(func = c("min", "min", "max", "min"), zone = c("BE", "FR", "DE", "DE")).
func is the direction of the fix faces and zone is the area of this direction.
If you give for example min and DE, there will be a fix face at the minimum import
value of Germany.}

\item{VERTRawDetails}{\code{data.table}, vertices of the polyhedron A}
}
\description{
This function takes in input two polyhedrons, the reference one and
the moving one (or set of facets). It returns in output a really close polyhedron,
to the reference one, with the set of facets. It runs a minimization program 
on a distance using the intersections of a set of lines and the two polyhedrons 
and validate the moving polyhedron with a volume indicator.
}
\examples{
\dontrun{
library(data.table)
library(quadprog)
library(linprog)
polyhedra <- readRDS(system.file("testdata/polyhedra.rds", package = "fbAntares"))
A <- polyhedra[Date == "2019-02-14"]
B <- polyhedra[Date == "2019-02-15"]
nbLines <- 10000
maxiter <- 20
thresholdIndic <- 0.9
quad <- F

 getBestPolyhedron(
 A = A, B = B, nbLines = nbLines, maxiter = maxiter,
 thresholdIndic = thresholdIndic, quad = quad)
}

}
