% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatsFBts.R
\name{getStatsFBts}
\alias{getStatsFBts}
\title{Return a list with statistics summary of the file ts.txt}
\usage{
getStatsFBts(ts, interSeasonBegin = c("2028-10-01", "2029-03-16"),
  interSeasonEnd = c("2028-10-31", "2029-05-15"), output = "summary")
}
\arguments{
\item{ts}{\code{data.table} The output ts obtained with \link{createFBTS}}

\item{interSeasonBegin}{\code{character or date}, date or vector of dates, YYYY-MM-DD, 
begin of interseason}

\item{interSeasonEnd}{\code{character or date}, date or vector of dates, YYYY-MM-DD, 
end of interseason}

\item{output}{\code{character}, the output of you want, either summary, 
yearbyyear or all, default is summary, if you want the details check the example.}
}
\description{
This function takes the object ts, generated with \link{createFBTS}
and return a list with statistics summary of this object
}
\examples{
\dontrun{
ts <- fread(system.file("testdata/antaresInput/ts.txt", package = "fbAntares"),
header = T)
interSeasonBegin <-  c("2028-10-01", "2029-03-16")
interSeasonEnd <- c("2028-10-31", "2029-05-15")
statsFBts <- getStatsFBts(ts, interSeasonBegin, interSeasonEnd, output = "summary")
statsFBts <- getStatsFBts(ts, interSeasonBegin, interSeasonEnd, output = "yearbyyear")
statsFBts <- getStatsFBts(ts, interSeasonBegin, interSeasonEnd, output = "all")
}
}
